Modified compute::delete recipe for openstack. (#21)
Previously we would determine the compute to delete by using its ID which is
unique. The problem is in debug mode we do not purge VMs that ran into issues
during add recipe, and we end up with multiple VMs with the same name but
different IDs. When we run delete recipe we would only delete one of such VMs,
leaving the rest orphaned. The new approach fixes that problem, by searching
the VMs by name, and running delete action for all found VMs.