Added usergroup to user component.
New functionality is added to user component to handle multiple users in one
component, by using a map - username-sshkeys. The checkbox attribute is added
to toggle between the standard way (one user per component) and the new way -
multiple users per component.
When doing multiple users only username and sshkeys will be specific for each
user, the rest of the attributes (ulimit, bash shell etc) will be the same for
all the users. The home directory attribute will be ignored and instead the
users will be using /home/username directory.
I purposefully didn't change indentations in the recipes, so the changes can
be minimal and it'll be a bit easier to review.
Will fix the indents in the next PR, along with any fixes we find during
testing.