Fixed a bug with azure eph disk, and a new KCI test for place_on_root. (#197)
There was a silent failure in azure deployments with xfs filesystem where the
first deployment would always fail, and the retry would succeed. The reason was
the xfsprogs package needed to be installed sooner in the code flow.
Also, we've recently added a new place_on_root functionality that would ignore
mounting disks in specific azure deployments, and instead just create a folder
on root. But we never adjusted KCI tests and once we added KCI to main-1 (where
that functionality was enabled by default for a few packs) the tests started to
fail. Fixing it by having a specific spec file just for place_on_root cases.