Ensure Managed Disk Deletion in Azure
Summary: Even after a managed disk has been deleted, it manages to show up in the 'list_resources' method (this method is used to list all the resources within a RG) for up to 2 minutes after delete is called on it. Because of this, the resource groups created were not getting deleted but remaining there as stale resources. This was causing RG quota issues on Azure. Code has been added to ensure that the disk is fully gone after its deletion, before mobing on to the next part of the code.